Use running.is_set() to check running flag
Check API version

Pass not decoded packets to handle functions so that if we need to
forward the packets we don't have to build them new. Specifically for
data packets w=this would keep signatures intact.

Pass address instead of node name so that this can be stored in PIT